backend/requirements.txt
asgiref==3.5.1
requests==2.27.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django==3.2.13
djangorestframework==3.13.1
djoser==2.1.0
django-filter==21.1
Pillow==9.1.0
django-import-export==2.8.0